requirements.txt
fastapi
uvicorn
torch
torchvision
transformers
pillow
opencv-python
numpy
jinja2


